Move NickSearch and BylineSearch to demoscene.fields and productions.fields
