+++ demoscene/fields/__init__.py
@@ -1,2 +1,3 @@
 from .matched_nick_field import MatchedNickField  # noqa
 from .nick_field import NickField, NickLookup  # noqa
+from .nick_search import NickSearch, NickSelection  # noqa
