+++ demoscene/fields/matched_nick_field.py
@@ -6,7 +6,8 @@ from django.utils.html import conditional_escape
 from django.utils.safestring import mark_safe
 
 from demoscene.models import Nick
-from demoscene.utils.nick_search import NickSelection
+
+from .nick_search import NickSelection
 
 
 class NickChoicesWidget(forms.RadioSelect):
